feat(config): texture combined
