+++ common/src/configs/mod.rs
@@ -3,17 +3,21 @@ pub mod model_config;
 pub mod physics_config;
 pub mod player_config;
 pub mod scene_config;
+pub mod texture_config;
 
-use crate::configs::audio_config::ConfigAudioAssets;
-use crate::configs::model_config::ConfigModels;
-use crate::configs::player_config::ConfigPlayer;
-use crate::configs::scene_config::ConfigSceneGraph;
 use once_cell::sync::Lazy as OnceCellLazy;
 use serde::Serialize;
 use std::fs::File;
 use std::io::{Read, Write};
 use std::path::Path;
 use std::sync::{Arc, RwLock};
+
+use crate::configs::audio_config::ConfigAudioAssets;
+use crate::configs::model_config::ConfigModels;
+use crate::configs::player_config::ConfigPlayer;
+use crate::configs::scene_config::ConfigSceneGraph;
+use crate::configs::texture_config::ConfigTexture;
+
 
 const MODELS_CONFIG_PATH: &str = "models.json";
 const SCENE_CONFIG_PATH: &str = "scene.json";
@@ -22,11 +26,16 @@ const PLAYER_CONFIG_PATH: &str = "player.json";
 const DISPLAY_CONFIG_PATH: &str = "display.json";
 const TEXTURE_CONFIG_PATH: &str = "tex.json";
 
+pub static CONFIG_INSTANCE: OnceCellLazy<RwLock<Option<Arc<Config>>>> =
+    OnceCellLazy::new(|| RwLock::new(None));
+
 pub struct Config {
     pub models: ConfigModels,
     pub scene: ConfigSceneGraph,
     pub audio: ConfigAudioAssets,
     pub player: ConfigPlayer,
+    //pub display: ConfigDisplay,
+    pub texture: ConfigTexture,
 }
 
 impl Config {
@@ -35,12 +44,16 @@ impl Config {
         scene: ConfigSceneGraph,
         audio: ConfigAudioAssets,
         player: ConfigPlayer,
+        //display: ConfigDisplay,
+        texture: ConfigTexture
     ) -> Self {
         Config {
             models,
             scene,
             audio,
             player,
+            //display,
+            texture,
         }
     }
 }
@@ -53,8 +66,10 @@ impl ConfigurationManager {
         let scene: ConfigSceneGraph = from_file(SCENE_CONFIG_PATH)?;
         let audio: ConfigAudioAssets = from_file(AUDIO_CONFIG_PATH)?;
         let player: ConfigPlayer = from_file(PLAYER_CONFIG_PATH)?;
+        //let display: ConfigDisplay = from_file(DISPLAY_CONFIG_PATH)?;
+        let texture: ConfigTexture = from_file(TEXTURE_CONFIG_PATH)?;
 
-        let config = Config::new(models, scene, audio, player);
+        let config = Config::new(models, scene, audio, player, texture);
         *CONFIG_INSTANCE.write().unwrap() = Some(Arc::new(config));
         Ok(())
     }
@@ -68,9 +83,6 @@ impl ConfigurationManager {
             .clone()
     }
 }
-
-pub static CONFIG_INSTANCE: OnceCellLazy<RwLock<Option<Arc<Config>>>> =
-    OnceCellLazy::new(|| RwLock::new(None));
 
 pub fn from_file<P: AsRef<Path>, S: serde::de::DeserializeOwned>(
     path: P,
